// ui/src/lib/api/generated/model/comment.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire API
 * Centoire community platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CommentStatus } from './commentStatus';
import type { CommentAuthor } from './commentAuthor';

export interface Comment {
  id: string;
  content: string;
  status: CommentStatus;
  depth: number;
  upvoteCount: number;
  replyCount: number;
  createdAt: string;
  /** @nullable */
  author: CommentAuthor;
  viewerVoted: boolean;
  replies: Comment[];
}